Normalize refactor workspace directory structure from engine/ to src/
