+++ src/refactor/tsconfig.json
@@ -15,6 +15,6 @@
     { "path": "../semantic" },
     { "path": "../plugin" }
   ],
-  "include": ["*.ts", "engine/**/*.ts", "test/**/*.ts"],
+  "include": ["*.ts", "src/**/*.ts", "test/**/*.ts"],
   "exclude": ["dist"]
 }
